corrected directory structure and uvm run scripts to conform to amamory requirements for git push

+++ run.sh
@@ -1,17 +1,39 @@
 #!/bin/bash
-module load modelsim
-echo "	simulating... please wait"
-{
-  #run vsim
 
-  vsim -c -do ./scripts/3_simul.do
+if [ "$1" = "sv" ]
+then
+
+  module load modelsim
+  echo "	simulating... please wait"
+  {
+    #run vsim
+
+    vsim -c -do ./scripts/3_simul.do
+
+    #save log file
+    mv transcript simulation.log
+
+    #cleanup
+    rm -rf *.ini transcript *.wlf work ./src/*~
+
+  } > /dev/null
+  module unload modelsim
+
+elif [ "$1" = "uvm" ]
+then
+
+  #run sim
+  vsim -c -do scripts/uvm/1_compile.do
 
   #save log file
   mv transcript simulation.log
 
   #cleanup
-  rm -rf *.ini transcript *.wlf work ./src/*~
+  rm -rf dump.vcd teste* transcript utopia.ucdb  vsim.wlf work
 
-} > /dev/null
-module unload modelsim
+else
+
+  echo "NOT AVAILABLE OPTION. Please, run \"./run.sh sv\" or \"./run.sh uvm\" "
+
+fi
